Creation methode utilitaire génération items calendrier
