+++ Client/TODO.txt
@@ -4,3 +4,4 @@
 [] Renommer les rotues ex : STUDENT_PROFILE, HOST_PROFILE
 [] Calendar component le meme dans le home du student que dans le host
 [] Unit test utils class
+[] Enlever les getters inutils des stores
