+++ Client/src/app/stores/StudentStore.ts
@@ -8,7 +8,7 @@
 import { action, makeAutoObservable, observable, runInAction } from 'mobx';
 import { Meeting, Location, Host, Chat, Message, Filter } from '../models/ApplicationTypes';
 import AmphitryonDAO from '../data/AmphitryonDAO';
-import { addDays, endOfDay, format, startOfDay } from 'date-fns';
+import { addDays, endOfDay, startOfDay } from 'date-fns';
 import { Alert } from 'react-native';
 import { AgendaItemsMap } from 'react-native-calendars';
 import Strings from '../context/Strings';
@@ -414,28 +414,7 @@ class StudentStore {
     await this.loadUserMeetings(startOfDay(from), endOfDay(addDays(from, 10)));
     runInAction(() => {
       this.dateInCalendar = from;
-      const nbDays = 10;
-      const items = ['{ '];
-      for (let i = 0; i <= nbDays; ++i) {
-        items.push('"' + format(addDays(from, i), 'yyyy-MM-dd') + '" : [');
-        const dayMeetings = this.userMeetings?.filter((current: Meeting) => {
-          return (
-            format(new Date(current.startDate), 'yyyy-MM-dd') ===
-            format(addDays(from, i), 'yyyy-MM-dd')
-          );
-        });
-
-        let cpt = 0;
-        dayMeetings?.map((current: Meeting) => {
-          items.push(JSON.stringify(current));
-          cpt++;
-          if (cpt !== dayMeetings?.length) items.push(',');
-        });
-        items.push(']');
-        if (i != nbDays) items.push(',');
-      }
-      items.push(' }');
-      this.items = JSON.parse(items.join(''));
+      this.items = this.utils.generateItems(this.userMeetings, from);
     });
   }
