+++ Client/src/app/utils/Utils.ts
@@ -5,8 +5,11 @@
  * @brief   Utils class
  */
 
+import { addDays, format } from 'date-fns';
 import { Alert } from 'react-native';
+import { AgendaItemsMap } from 'react-native-calendars';
 import Strings from '../context/Strings';
+import { Meeting } from '../models/ApplicationTypes';
 
 class Utils {
   private static instance: Utils;
@@ -50,6 +53,37 @@ class Utils {
       default:
         Alert.alert(Strings.ERROR_OCCURED, Strings.ERROR_UNEXPECTED);
     }
+  }
+
+  /**
+   * Generated items for calendar view with meetings
+   * @param meetings to generate items
+   * @param from date from to filter meetings to display
+   * @returns items
+   */
+  generateItems(meetings: Meeting[], from: Date): AgendaItemsMap<Meeting> {
+    const nbDays = 10;
+    const items = ['{ '];
+    for (let i = 0; i <= nbDays; ++i) {
+      items.push('"' + format(addDays(from, i), 'yyyy-MM-dd') + '" : [');
+      const dayMeetings = meetings.filter((current: Meeting) => {
+        return (
+          format(new Date(current.startDate), 'yyyy-MM-dd') ===
+          format(addDays(from, i), 'yyyy-MM-dd')
+        );
+      });
+
+      let cpt = 0;
+      dayMeetings?.map((current: Meeting) => {
+        items.push(JSON.stringify(current));
+        cpt++;
+        if (cpt !== dayMeetings?.length) items.push(',');
+      });
+      items.push(']');
+      if (i != nbDays) items.push(',');
+    }
+    items.push(' }');
+    return JSON.parse(items.join(''));
   }
 }
